Refactor, move tagger related functions/constants from the corpus submodules (lst20/orchid) to tagger submodules (unigram/perceptron).
