+++ pythainlp/tag/orchid.py
@@ -2,8 +2,9 @@
 """
 Data preprocessing for ORCHID corpus
 """
-from typing import List
+from typing import List, Tuple
 
+# ORCHID part-of-speech tags
 ORCHID_SIGN_TAGS = {
     " ": "<space>",
     "+": "<plus>",
@@ -31,10 +32,105 @@ ORCHID_SIGN_TAGS = {
 }
 ORCHID_SIGN_TEXTS = dict((v, k) for k, v in ORCHID_SIGN_TAGS.items())
 
+# map from ORCHID POS tag to Universal POS tag
+# from Korakot Chaovavanich
+ORCHID_TO_UD = {
+    # NOUN
+    "NOUN": "NOUN",
+    "NCMN": "NOUN",
+    "NTTL": "NOUN",
+    "CNIT": "NOUN",
+    "CLTV": "NOUN",
+    "CMTR": "NOUN",
+    "CFQC": "NOUN",
+    "CVBL": "NOUN",
+    # VERB
+    "VACT": "VERB",
+    "VSTA": "VERB",
+    # PROPN
+    "PROPN": "PROPN",
+    "NPRP": "PROPN",
+    # ADJ
+    "ADJ": "ADJ",
+    "NONM": "ADJ",
+    "VATT": "ADJ",
+    "DONM": "ADJ",
+    # ADV
+    "ADV": "ADV",
+    "ADVN": "ADV",
+    "ADVI": "ADV",
+    "ADVP": "ADV",
+    "ADVS": "ADV",
+    # INT
+    "INT": "INTJ",
+    # PRON
+    "PRON": "PRON",
+    "PPRS": "PRON",
+    "PDMN": "PRON",
+    "PNTR": "PRON",
+    # DET
+    "DET": "DET",
+    "DDAN": "DET",
+    "DDAC": "DET",
+    "DDBQ": "DET",
+    "DDAQ": "DET",
+    "DIAC": "DET",
+    "DIBQ": "DET",
+    "DIAQ": "DET",
+    # NUM
+    "NUM": "NUM",
+    "NCNM": "NUM",
+    "NLBL": "NUM",
+    "DCNM": "NUM",
+    # AUX
+    "AUX": "AUX",
+    "XVBM": "AUX",
+    "XVAM": "AUX",
+    "XVMM": "AUX",
+    "XVBB": "AUX",
+    "XVAE": "AUX",
+    # ADP
+    "ADP": "ADP",
+    "RPRE": "ADP",
+    # CCONJ
+    "CCONJ": "CCONJ",
+    "JCRG": "CCONJ",
+    # SCONJ
+    "SCONJ": "SCONJ",
+    "PREL": "SCONJ",
+    "JSBR": "SCONJ",
+    "JCMP": "SCONJ",
+    # PART
+    "PART": "PART",
+    "FIXN": "PART",
+    "FIXV": "PART",
+    "EAFF": "PART",
+    "EITT": "PART",
+    "AITT": "PART",
+    "NEG": "PART",
+    # PUNCT
+    "PUNCT": "PUNCT",
+    "PUNC": "PUNCT",
+}
+
+
+def ud_exception(w: str, tag: str) -> str:
+    if w == "การ" or w == "ความ":
+        return "NOUN"
+
+    return tag
+
+
+def to_ud(word_tags: List[Tuple[str, str]]) -> List[Tuple[str, str]]:
+    return [
+        (word_tag[0], ud_exception(word_tag[0], ORCHID_TO_UD[word_tag[1]]))
+        for word_tag in word_tags
+    ]
+
 
 def tag_signs(words: List[str]) -> List[str]:
     """
-    Tag signs and symbols with their tags.
+    Mark signs and symbols with their tags.
     This function is to be used a preprocessing before the actual POS tagging.
     """
     keys = ORCHID_SIGN_TAGS.keys()
