+++ pythainlp/tag/unigram.py
@@ -6,18 +6,21 @@ import json
 import os
 from typing import List, Tuple
 
-from pythainlp.corpus import corpus_path
-from pythainlp.tag.orchid import tag_signs, tag_to_text
-from pythainlp.tag.lst20 import (
-    _lst20_tagger,
-    lst20_tag_signs,
-    lst20_tag_to_text,
-)
+from pythainlp.corpus import corpus_path, get_corpus_path
 
-_THAI_POS_ORCHID_FILENAME = "orchid_pos_th.json"
-_THAI_POS_ORCHID_PATH = os.path.join(corpus_path(), _THAI_POS_ORCHID_FILENAME)
-_THAI_POS_PUD_FILENAME = "ud_thai_pud_unigram_tagger.json"
-_THAI_POS_PUD_PATH = os.path.join(corpus_path(), _THAI_POS_PUD_FILENAME)
+
+_ORCHID_FILENAME = "orchid_pos_th.json"
+_ORCHID_PATH = os.path.join(corpus_path(), _ORCHID_FILENAME)
+
+_PUD_FILENAME = "ud_thai_pud_unigram_tagger.json"
+_PUD_PATH = os.path.join(corpus_path(), _PUD_FILENAME)
+
+_LST20_TAGGER_NAME = "lst20_unigram_tagger"
+
+
+_ORCHID_TAGGER = None
+_PUD_TAGGER = None
+_LST20_TAGGER = None
 
 
 def _find_tag(words: List[str], dictdata: dict) -> List[Tuple[str, str]]:
@@ -29,21 +32,35 @@ def _find_tag(words: List[str], dictdata: dict) -> List[Tuple[str, str]]:
 
 
 def _orchid_tagger():
-    with open(_THAI_POS_ORCHID_PATH, encoding="utf-8-sig") as f:
-        model = json.load(f)
-    return model
+    global _ORCHID_TAGGER
+    if not _ORCHID_TAGGER:
+        with open(_ORCHID_PATH, encoding="utf-8-sig") as fh:
+            _ORCHID_TAGGER = json.load(fh)
+    return _ORCHID_TAGGER
 
 
 def _pud_tagger():
-    with open(_THAI_POS_PUD_PATH, encoding="utf-8-sig") as f:
-        model = json.load(f)
-    return model
+    global _PUD_TAGGER
+    if not _PUD_TAGGER:
+        with open(_PUD_PATH, encoding="utf-8-sig") as fh:
+            _PUD_TAGGER = json.load(fh)
+    return _PUD_TAGGER
 
 
-def _postag_clean(words, tagger, tag_sign, to_text):
-    words = tag_sign(words)
+def _lst20_tagger():
+    global _LST20_TAGGER
+    if not _LST20_TAGGER:
+        path = get_corpus_path(_LST20_TAGGER_NAME)
+        with open(path, encoding="utf-8-sig") as fh:
+            _LST20_TAGGER = json.load(fh)
+    return _LST20_TAGGER
+
+
+# actual tagging work happens here
+def _postag_clean(words: List[str], tagger, tag_signs, to_text):
+    words = tag_signs(words)
     word_tags = _find_tag(words, tagger())
-    return [(tag_to_text(word_tag[0]), word_tag[1]) for word_tag in word_tags]
+    return [(to_text(word_tag[0]), word_tag[1]) for word_tag in word_tags]
 
 
 def tag(words: List[str], corpus: str) -> List[Tuple[str, str]]:
@@ -54,14 +71,20 @@ def tag(words: List[str], corpus: str) -> List[Tuple[str, str]]:
         return []
 
     word_tags = []
-    if corpus == "orchid":
+    if corpus == "orchid" or corpus == "orchid_ud":
+        from pythainlp.tag.orchid import tag_signs, tag_to_text, to_ud
+
         word_tags = _postag_clean(
             words, _orchid_tagger, tag_signs, tag_to_text
         )
-    elif corpus == "lst20":
-        word_tags = _postag_clean(
-            words, _lst20_tagger, lst20_tag_signs, lst20_tag_to_text
-        )
+        if corpus == "orchid_ud":
+            word_tags = to_ud(word_tags)
+    elif corpus == "lst20" or corpus == "lst20_ud":
+        from pythainlp.tag.lst20 import tag_signs, tag_to_text, to_ud
+
+        word_tags = _postag_clean(words, _lst20_tagger, tag_signs, tag_to_text)
+        if corpus == "lst20_ud":
+            word_tags = to_ud(word_tags)
     else:
         word_tags = _find_tag(words, _pud_tagger())
